+++ test/unit_testing/tests/NoMainFunc/intended/include/HxArray.h
@@ -18,7 +18,7 @@ public:
 			{
 				int _g = 0;
 				std::shared_ptr<std::deque<T>> _g1 = a;
-				while(_g < _g1->size()) {
+				while((unsigned int)(_g) < _g1->size()) {
 					T obj = (*_g1)[_g];
 					
 					++_g;
@@ -32,7 +32,7 @@ public:
 		
 		{
 			int _g = 0;
-			while(_g < other->size()) {
+			while((unsigned int)(_g) < other->size()) {
 				T o = (*other)[_g];
 				
 				++_g;
@@ -67,7 +67,7 @@ public:
 		if(pos < 0) {
 			pos += a->size();
 		};
-		if(pos < 0 || pos >= a->size()) {
+		if(pos < 0 || (unsigned int)(pos) >= a->size()) {
 			return std::make_shared<std::deque<T>>(std::deque<T>{});
 		};
 		if(!end.has_value() || end.value() > a->size()) {
@@ -88,7 +88,7 @@ public:
 		while(_g < _g1.value()) {
 			int i = _g++;
 			
-			if(i >= 0 && i < a->size()) {
+			if(i >= 0 && (unsigned int)(i) < a->size()) {
 				result->push_back((*a)[i]);
 			};
 		};
@@ -101,13 +101,13 @@ public:
 		if(pos < 0) {
 			pos += a->size();
 		};
-		if(pos < 0 || pos > a->size()) {
+		if(pos < 0 || (unsigned int)(pos) > a->size()) {
 			return std::make_shared<std::deque<T>>(std::deque<T>{});
 		};
 		if(len < 0) {
 			return std::make_shared<std::deque<T>>(std::deque<T>{});
 		};
-		if(pos + len > a->size()) {
+		if((unsigned int)(pos + len) > a->size()) {
 			len = a->size() - pos;
 		};
 		
@@ -121,7 +121,7 @@ public:
 		while(_g < _g1) {
 			int i = _g++;
 			
-			if(i >= 0 && i < a->size()) {
+			if(i >= 0 && (unsigned int)(i) < a->size()) {
 				result->push_back((*a)[i]);
 			};
 		};
@@ -163,7 +163,7 @@ public:
 		std::shared_ptr<std::deque<S>> _g = std::make_shared<std::deque<S>>(std::deque<S>{});
 		int _g1 = 0;
 		
-		while(_g1 < a->size()) {
+		while((unsigned int)(_g1) < a->size()) {
 			T v = (*a)[_g1];
 			
 			++_g1;
@@ -180,7 +180,7 @@ public:
 		std::shared_ptr<std::deque<T>> _g = std::make_shared<std::deque<T>>(std::deque<T>{});
 		int _g1 = 0;
 		
-		while(_g1 < a->size()) {
+		while((unsigned int)(_g1) < a->size()) {
 			T v = (*a)[_g1];
 			
 			++_g1;
