+++ test/unit_testing/tests/Sys/intended-Windows/src/Main.cpp
@@ -131,7 +131,7 @@ void Main::main() {
 	
 	std::shared_ptr<std::deque<std::string>> args = Sys_Args::args();
 	
-	Main::assert(args->size() == 1, haxe::shared_anon<haxe::PosInfos>("Main"s, "test/unit_testing/tests/Sys/Main.hx"s, 33, "main"s));
+	Main::assert(args->size() == (unsigned int)(1), haxe::shared_anon<haxe::PosInfos>("Main"s, "test/unit_testing/tests/Sys/Main.hx"s, 33, "main"s));
 	Main::assert((*args)[0].find("test_out"s) >= 0, haxe::shared_anon<haxe::PosInfos>("Main"s, "test/unit_testing/tests/Sys/Main.hx"s, 34, "main"s));
 	haxe::Log::trace(std::filesystem::current_path().string(), haxe::shared_anon<haxe::PosInfos>("Main"s, "test/unit_testing/tests/Sys/Main.hx"s, 38, "main"s));
 	std::filesystem::current_path("C:/"s);
